[FIX] Fix sonar issues
